add strict-build-order option to build
Adds an option to enforce strict build ordering when building packages.
In strict ordering the studio package used to build a package is
considered as dependency of the package because the studio environment
may affect the actual build process though it ideally should not.

When strict ordering is not enforced we assume that any update to the
studio does not actually have an impact on the build of a package. This
is useful when refreshing packages.

Signed-off-by: Johny Jose <johny.jose@progress.com>