removes prompts and fixes studio build order
You can now pass -n to the build command to prevent the prompt from
appearing, asking whether you want to skip certain packages.
However, please note that if a '.hab-build-ignore' file is present, it will
still affect the build order of the packages and any packages specified
in that file will be skipped during the build process.

Signed-off-by: Johny Jose <johny.jose@progress.com>